+++ extension.ts
@@ -747,10 +747,11 @@ export function activate(context: vscode.ExtensionContext) {
 
 	// function used to attach bookmarks at the line
     function stickyBookmarks(event): boolean {
-        let useStickyBookmarks: boolean = vscode.workspace.getConfiguration("bookmarks").get("useStickyBookmarks", false);
-        if (!useStickyBookmarks) {
-            return false;
-        }
+        // sticky is now the default/only behavior
+        // let useStickyBookmarks: boolean = vscode.workspace.getConfiguration("bookmarks").get("useStickyBookmarks", false);
+        // if (!useStickyBookmarks) {
+        //     return false;
+        // }
 
         let diffLine: number;
         let updatedBookmark: boolean = false;