+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 0.13.0
 
 * **New Setting:** Bookmarks can now be saved in the project (inside `.vscode` folder)
+* **Removed Setting:** Bookmarks are now _always_ Sticky
 
 ## Version 0.12.0